requirements.txt
pydantic
requests
python-dotenv
tqdm
fastapi
uvicorn
streamlit
imapclient
pyzmail36
selenium
beautifulsoup4
streamlit-authenticator==0.1.5
sqlalchemy
sqlalchemy-utils
bcrypt
xlsxwriter
reportlab
python-multipart
capsolver

